# docs/openapi-schema.yml
openapi: 3.0.2
info:
  title: 'Book DB'
  version: ''
paths:
  /api/users/:
    get:
      operationId: List of users
      description: Endpoint is avaliable for all. Avaliable response with authors and books
      parameters:
        - name: with_authors
          required: false
          in: query
          description: Show users with created authors.
          schema:
            type: bool
        - name: with_books
          required: false
          in: query
          description: Show users with created books.
          schema:
            type: bool
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
      tags:
        - Users
    post:
      operationId: Sing up
      description: ''
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomUserCreate'
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
          description: 'Пользователь успешно создан'
        '400':
          $ref: '#/components/responses/ValidationError'
      tags:
        - Users

  /api/users/{id}/:
    get:
      operationId: User profile
      description: 'Available to all users.'
      security:
        - Token: [ ]
      parameters:
        - name: id
          in: path
          required: true
          description: "Unique ID of the user"
          schema:
            type: int
        - name: with_authors
          required: false
          in: query
          description: Show users with created authors.
          schema:
            type: bool
        - name: with_books
          required: false
          in: query
          description: Show users with created books.
          schema:
            type: bool
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
          description: ''
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/AuthenticationError'
      tags:
      - Users
    put:
      operationId: Update user
      description: 'Available to all users.'
      security:
        - Token: [ ]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: int
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomUserCreate'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
          description: ''
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/AuthenticationError'
      tags:
      - Users

  /api/auth/:
    post:
      operationId: Get token
      description: Used for authorization by username and password, in order to further use the token for requests.
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TokenCreate'
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenGetResponse'
          description: ''
      tags:
        - Users

  /api/books/:
    get:
      operationId: List of books
      description: Endpoint is avaliable for all. Avaliable response with author
      parameters:
        - name: with_authors
          required: false
          in: query
          description: Show book with author.
          schema:
            type: bool
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Book'
      tags:
        - Books
    post:
      security:
        - Token: []
      operationId: Create book
      description: 'Avaliable only for authorized'
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookCreate'
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
          description: 'Book created successfully'
        '400':
          description: 'Standart errors'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '401':
          $ref: '#/components/schemas/AuthenticationError'
        '404':
          $ref: '#/components/responses/NotFound'
      tags:
        - Books

  /api/books/{id}/:
    get:
      operationId: Get book
      description: ''
      parameters:
        - name: id
          in: path
          required: true
          description: "Unique ID of book"
          schema:
            type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
          description: ''
      tags:
        - Books
    put:
      operationId: Update book
      security:
        - Token: [ ]
      description: 'Avaliable only to book owner'
      parameters:
        - name: id
          in: path
          required: true
          description: "Unique ID of the book."
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookCreate'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
          description: 'Book updated successfully'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/AuthenticationError'
        '403':
          $ref: '#/components/responses/PermissionDenied'
        '404':
          $ref: '#/components/responses/NotFound'
      tags:
        - Books
    delete:
      operationId: Delete book

      description: 'Avaliable only to book owner'
      security:
        - Token: [ ]
      parameters:
        - name: id
          in: path
          required: true
          description: "Unique ID of the book"
          schema:
            type: string
      responses:
        '204':
          description: 'Book deleted successfully'
        '401':
          $ref: '#/components/responses/AuthenticationError'
        '403':
          $ref: '#/components/responses/PermissionDenied'
        '404':
          $ref: '#/components/responses/NotFound'
      tags:
        - Books

  /api/authors/:
    get:
      operationId: List of authors
      description: Endpoint is avaliable for all.
      parameters: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Author'
      tags:
        - Authors
    post:
      security:
        - Token: []
      operationId: Create author
      description: 'Avaliable only for authorized'
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Author'
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
          description: 'Book created successfully'
        '400':
          description: 'Standart errors'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '401':
          $ref: '#/components/schemas/AuthenticationError'
        '404':
          $ref: '#/components/responses/NotFound'
      tags:
        - Authors

  /api/authors/{id}/:
    get:
      operationId: Get Author
      description: ''
      parameters:
        - name: id
          in: path
          required: true
          description: "Unique ID of author"
          schema:
            type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Author'
          description: ''
      tags:
        - Authors
    put:
      operationId: Update author
      security:
        - Token: [ ]
      description: 'Avaliable only to author owner'
      parameters:
        - name: id
          in: path
          required: true
          description: "Unique ID of the authir."
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Author'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Author'
          description: 'Authors updated successfully'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/AuthenticationError'
        '403':
          $ref: '#/components/responses/PermissionDenied'
        '404':
          $ref: '#/components/responses/NotFound'
      tags:
        - Authors
    delete:
      operationId: Delete author

      description: 'Avaliable only to author owner'
      security:
        - Token: [ ]
      parameters:
        - name: id
          in: path
          required: true
          description: "Unique ID of the author"
          schema:
            type: string
      responses:
        '204':
          description: 'Authors deleted successfully'
        '401':
          $ref: '#/components/responses/AuthenticationError'
        '403':
          $ref: '#/components/responses/PermissionDenied'
        '404':
          $ref: '#/components/responses/NotFound'
      tags:
        - Authors

components:
  schemas:
    User:
      description:  'User (In aithors and book - creator)'
      type: object
      properties:
        email:
          type: string
          format: email
          maxLength: 120
          description: "Email"
        id:
          type: integer
          readOnly: true
        username:
          type: string
          description: "Unique username"
          pattern: ^[\w.@+-]+\z
          maxLength: 50
      required:
        - username

    Book:
      description:  'Book'
      type: object
      properties:
        created:
          type: string
          format: date
          example: "2022-02-06 11:33:06"
          readOnly: true
        id:
          type: integer
          readOnly: true
        isbn:
          type: integer
        number_of_pages:
          type: integer
        review:
          type: string
          description: "Review"
        title:
          type: string
          description: "Title"
          maxLength: 120
      required:
        - isbn
        - number_of_pages
        - review
        - title

    BookCreate:
      description:  'Create Book'
      type: object
      properties:
        created:
          type: string
          format: date
          example: "2022-02-06 11:33:06"
          readOnly: true
        author:
          type: integer
        isbn:
          type: integer
        number_of_pages:
          type: integer
        review:
          type: string
          description: "Review"
        title:
          type: string
          description: "Title"
          maxLength: 120
      required:
        - isbn
        - number_of_pages
        - review
        - title

    Author:
      description:  'BAuthorook'
      type: object
      properties:
        created:
          type: string
          format: date
          example: "2022-02-06 11:33:06"
          readOnly: true
        id:
          type: integer
          readOnly: true
        isbn:
          type: integer
        fio:
          type: string
          description: "FIO"
          maxLength: 120
      required:
        - isbn
        - fio

    CustomUserCreate:
      type: object
      properties:
        email:
          type: string
          format: email
          maxLength: 120
          description: "Email"
          example: "vpupkin@yandex.ru"
        id:
          type: integer
          readOnly: true
        username:
          type: string
          description: "Unique username"
          pattern: ^[\w.@+-]+\z
          maxLength: 50
          example: "vasya.pupkin"
        password:
          type: string
          maxLength: 120
          description: "Password"
          example: "Qwerty123"
      required:
        - username
        - password
        - email
    Activation:
      type: object
      properties:
        uid:
          type: string
        token:
          type: string
      required:
        - uid
        - token

    PasswordResetConfirm:
      type: object
      properties:
        uid:
          type: string
        token:
          type: string
        new_password:
          type: string
      required:
        - uid
        - token
        - new_password
    UsernameResetConfirm:
      type: object
      properties:
        new_email:
          type: string
          format: email
          maxLength: 254

    SetUsername:
      type: object
      properties:
        current_password:
          type: string
        new_email:
          type: string
          format: email
          maxLength: 254
      required:
        - current_password
    TokenCreate:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
    TokenGetResponse:
      type: object
      properties:
        auth_token:
          type: string
    RecipeCreateUpdate:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        ingredients:
          description: Список ингредиентов
          type: array
          items:
            example:
              id: 1123
              amount: 10
            type: object
            properties:
              id:
                description: 'Уникальный id'
                type: integer
              amount:
                description: 'Количество в рецепте'
                type: integer
            required:
              - id
              - amount
        tags:
          description: 'Список id тегов'
          type: array
          example: [1, 2]
          items:
            type: integer
        image:
          description: 'Картинка, закодированная в Base64'
          example: 'data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAAEAAAABAgMAAABieywaAAAACVBMVEUAAAD///9fX1/S0ecCAAAACXBIWXMAAA7EAAAOxAGVKw4bAAAACklEQVQImWNoAAAAggCByxOyYQAAAABJRU5ErkJggg=='
          type: string
          format: binary
        name:
          description: 'Название'
          type: string
          maxLength: 200
        text:
          description: 'Описание'
          type: string
        cooking_time:
          description: 'Время приготовления (в минутах)'
          type: integer
          minimum: 1
      required:
        - ingredients
        - tags
        - image
        - name
        - text
        - cooking_time

    ValidationError:
      description: Standart errors
      type: object
      properties:
        field_name:
          description: "Name of the field with error. There're can be several of them"
          example: [ 'Field is required.' ]
          type: array
          items:
            type: string

    AuthenticationError:
      description: User is not authorized
      type: object
      properties:
        description:
          description: 'Request does not contain an access token'
          example: "Request does not contain an access token."
          type: string
        error:
          description: 'Authorization Required'
          example: "Authorization Required."
          type: string
        status_code:
          description: 'Status code'
          example: 401
          type: int

    PermissionDenied:
      description: Not enough rights
      type: object
      properties:
        detail:
          description: 'Error description'
          example: "You do not have sufficient rights to perform this action."
          type: string
    NotFound:
      description: Object not found
      type: object
      properties:
        error:
          description: 'Error description'
          example: "Not found."
          type: string

  responses:
    ValidationError:
      description: 'Standart errors'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'

    AuthenticationError:
      description: User is not authorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/AuthenticationError'

    PermissionDenied:
      description: Not enough rights
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/PermissionDenied'

    NotFound:
      description: Object not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/NotFound'


  securitySchemes:
    Token:
      description: 'Token authorization. <br>
      !All requests on behalf of the user must be made with the header "Authorization: JWT TOKENVALUE"'
      type: http
      scheme: token